+++ scripts/codegen/main.go
@@ -1,0 +1,7 @@
+package main
+
+import "github.com/launchdarkly/terraform-provider-launchdarkly/scripts/codegen/cmd"
+
+func main() {
+	cmd.Execute()
+}