+++ CHANGELOG.md
@@ -4,6 +4,10 @@ ENHANCEMENTS:
 
 - Added the `hide_member_details` argument to the Datadog `config` for the `launchdarkly_audit_log_subscription` resource.  When `hide_member_details` is `true`, LaunchDarkly member information will be redacted before events are sent to Datadog.
 
+NOTES:
+
+- Added a callout to the `bypassRequiredApproval` action in documentation.
+
 ## [2.5.0] (February 7, 2022)
 
 FEATURES: